requirements.txt
ollama==0.4.4
requests==2.32.3
streamlit==1.40.0
streamlit-folium==0.23.0
folium==0.18.0
pandas==2.2.3
python-dotenv==1.0.1
pytest==8.3.3
h3==4.1.2
geopandas==1.0.1
shapely==2.0.6
beautifulsoup4==4.12.3
google-generativeai==0.8.3
